chore: refactored cleanup tests for simplicity
